+++ src/generate/BothMoves.h
@@ -17,9 +17,9 @@ public:
         static void generate_reverse(const Position<Board>&, Stack&);
         static void generate_promotions(const Position<Board>&, Stack&);
 
-        static size_t count(const Position<Board>&);
-        static size_t count_reverse(const Position<Board>&);
-        static size_t count_promotions(const Position<Board>&);
+        static int count(const Position<Board>&);
+        static int count_reverse(const Position<Board>&);
+        static int count_promotions(const Position<Board>&);
         
         static bool detect(const Position<Board>&);
         static bool detect_reverse(const Position<Board>&);
