+++ src/generate/Driver.hpp
@@ -25,21 +25,21 @@ void Driver<Color, Material, MoveType, Rules, Board>::generate_promotions(const 
 }
 
 template<bool Color, int Material, int MoveType, typename Rules, typename Board>
-size_t Driver<Color, Material, MoveType, Rules, Board>::count(const Position<Board>&) 
+int Driver<Color, Material, MoveType, Rules, Board>::count(const Position<Board>&) 
 { 
         static_assert(Material == Material::NONE, "Template function already defined.");
         return 0; 
 }
         
 template<bool Color, int Material, int MoveType, typename Rules, typename Board>
-size_t Driver<Color, Material, MoveType, Rules, Board>::count_reverse(const Position<Board>&) 
+int Driver<Color, Material, MoveType, Rules, Board>::count_reverse(const Position<Board>&) 
 { 
         static_assert(Material == Material::NONE, "Template function already defined.");
         return 0; 
 }
 
 template<bool Color, int Material, int MoveType, typename Rules, typename Board>
-size_t Driver<Color, Material, MoveType, Rules, Board>::count_promotions(const Position<Board>&) 
+int Driver<Color, Material, MoveType, Rules, Board>::count_promotions(const Position<Board>&) 
 { 
         static_assert(Material == Material::NONE, "Template function already defined.");
         return 0;
